+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Insert `arpeg` with `@order="up"` per default (or `@order="down"` with `CTRL + A`)
 * Insert control elements with @staff attribute of selected elements (fix for #12)
 * Fix trailing space in `arpeg@plist`(fix for #13)
+* Support for inserting supplied elements around artic/accid attributes (that get converted to elements) (fix for #14)
 * Do not save and insert slur below with `CTRL + S` on Windows/Linux (fix for #15)
 * Support for `@place="between"` on control elements (fix for #16)